+++ Puzzle.hpp
@@ -34,7 +34,7 @@ class Puzzle {
         void        addToList(Node &src);
         void        setGoal();
         void        calculateManhattan(Node &n);
-        std::unordered_map<int, int>&    getClosedList();
+        std::unordered_map<uint64_t, int>&    getClosedList();
         std::priority_queue<Node*, std::vector<Node*>, CompareF >&     getOpenList();
 
 
@@ -44,7 +44,7 @@ class Puzzle {
 
         int _size;
         std::priority_queue<Node*, std::vector<Node*>, CompareF >  _openlist;
-        std::unordered_map<int, int>  _closedlist;
+        std::unordered_map<uint64_t, int>  _closedlist;
         std::unordered_map<int, xy> _mapGoal;
 };
